requirements/base.txt
Django==4.2
django-environ==0.9.0
psycopg2-binary==2.9.5
djangorestframework==3.14.0

celery==5.3.5
django-celery-beat==2.5.0
django-celery-results==2.4.0

whitenoise==6.2.0

django-filter==22.1
django-extensions==3.2.1
django-cors-headers==3.13.0
django-storages==1.13.1

drf-jwt==1.19.2

boto3==1.24.71
attrs==22.1.0

djangorestframework-simplejwt==5.2.2
drf-spectacular==0.24.2

django-redis==5.2.0
Faker==15.1.1
factory-boy==3.2.1
pytest==7.2.0
pytest-django==4.5.2

django-treebeard==4.7
ippanel==2.0.7
pillow==10.2.0
